# resources/docs/swagger.yaml
definitions:
  request.ArticleCreateRequest:
    properties:
      articleContent:
        type: string
      articleTag:
        type: string
      articleTitle:
        type: string
      remark:
        type: string
      summary:
        type: string
    required:
    - articleContent
    - articleTag
    - articleTitle
    - summary
    type: object
  request.ArticleUpdateRequest:
    properties:
      articleContent:
        type: string
      articleTag:
        type: string
      articleTitle:
        type: string
      id:
        type: integer
      remark:
        type: string
      status:
        type: string
      summary:
        type: string
    required:
    - articleContent
    - articleTag
    - articleTitle
    - id
    - status
    - summary
    type: object
  request.DeptCreateRequest:
    properties:
      ancestors:
        description: 祖籍
        type: string
      deptCode:
        description: 部门编码
        type: string
      deptName:
        description: 部门名称
        type: string
      email:
        description: 邮箱
        type: string
      leader:
        description: 负责人
        type: string
      orderNum:
        description: 显示顺序
        minimum: 0
        type: integer
      parentID:
        description: 父部门id
        minimum: 0
        type: integer
      phone:
        description: 联系电话
        type: string
      remark:
        description: 备注
        type: string
    required:
    - deptCode
    - deptName
    type: object
  request.DeptUpdateRequest:
    properties:
      ancestors:
        description: 祖籍
        type: string
      deptCode:
        description: 部门编码
        type: string
      deptName:
        description: 部门名称
        type: string
      email:
        description: 邮箱
        type: string
      id:
        description: 部门id
        type: integer
      leader:
        description: 负责人
        type: string
      orderNum:
        description: 显示顺序
        minimum: 0
        type: integer
      parentID:
        description: 父部门id
        minimum: 0
        type: integer
      phone:
        description: 联系电话
        type: string
      remark:
        description: 备注
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
    required:
    - deptCode
    - deptName
    - id
    - status
    type: object
...
  request.SysConfigCreateRequest:
    properties:
      configKey:
        description: 参数键名
        type: string
      configName:
        description: 参数名称
        type: string
      configValue:
        description: 参数键值
        type: string
      remark:
        description: 备注
        type: string
    required:
    - configKey
    - configName
    - configValue
    type: object
  request.SysConfigUpdateRequest:
    properties:
      configName:
        description: 参数名称
        type: string
      configValue:
        description: 参数键值
        type: string
      id:
        description: 参数主键
        type: integer
      remark:
        description: 备注
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
    required:
    - configName
    - configValue
    - id
    type: object
...
  response.ArticleResponse:
    properties:
      articleContent:
        type: string
      articleTag:
        type: string
      articleTitle:
        type: string
      createBy:
        type: string
      createTime:
        type: string
      delFlag:
        type: string
      id:
        type: integer
      remark:
        type: string
      status:
        type: string
      summary:
        type: string
      updateBy:
        type: string
      updateTime:
        type: string
    type: object
  response.DeptResponse:
    properties:
      ancestors:
        description: 祖级列表
        type: string
      children:
        description: 字部门列表
        items:
          $ref: '#/definitions/response.DeptResponse'
        type: array
      childrenSize:
        description: 子部门数量
        type: integer
      createBy:
        description: 创建者
        type: string
      createTime:
        description: 创建时间
        type: string
      delFlag:
        description: 删除标志（0代表存在 2代表删除）
        type: string
      deptCode:
        description: 部门编码
        type: string
      deptName:
        description: 部门名称
        type: string
      id:
        description: 部门id
        type: integer
      orderNum:
        description: 显示顺序
        type: integer
      parentID:
        description: 父部门id
        type: integer
      remark:
        description: 备注
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
      updateBy:
        description: 更新者
        type: string
      updateTime:
        description: 更新时间
        type: string
    type: object
...
  response.RestRes:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: status ok
        type: string
    type: object
...
  response.SysConfigResponse:
    properties:
      configKey:
        description: 参数键名
        type: string
      configName:
        description: 参数名称
        type: string
      configValue:
        description: 参数键值
        type: string
      createBy:
        description: 创建者
        type: string
      createTime:
        description: 创建时间
        type: string
      delFlag:
        description: 删除标志（0代表存在 2代表删除）
        type: string
      id:
        description: 参数主键
        type: integer
      remark:
        description: 备注
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
      updateBy:
        description: 更新者
        type: string
      updateTime:
        description: 更新时间
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample golang project
  title: Admin GOGOGO
  version: 0.0.1
paths:
...
  /article:
    post:
      consumes:
      - application/json
      description: 文章文稿新增信息
      parameters:
      - description: 新增文章文稿信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.ArticleCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 文章文稿新增
      tags:
      - 文章文稿
    put:
      consumes:
      - application/json
      description: 文章文稿更新信息
      parameters:
      - description: 更新文章文稿信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.ArticleUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 文章文稿更新
      tags:
      - 文章文稿
  /article/:id:
    delete:
      consumes:
      - application/json
      description: 文章文稿信息伪删除
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 文章文稿伪删除
      tags:
      - 文章文稿
    get:
      consumes:
      - application/json
      description: 文章文稿详情信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.ArticleResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 文章文稿详情
      tags:
      - 文章文稿
  /article/p:
    post:
      consumes:
      - application/json
      description: 分页获取所有文章文稿信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.ArticleResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 文章文稿分页查询
      tags:
      - 文章文稿
  /config:
    post:
      consumes:
      - application/json
      description: 新增参数配置信息
      parameters:
      - description: 新增参数配置信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.SysConfigCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 参数配置新增
      tags:
      - 系统参数配置
    put:
      consumes:
      - application/json
      description: 更新参数配置信息
      parameters:
      - description: 更新参数配置信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.SysConfigUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 参数配置更新
      tags:
      - 系统参数配置
  /config/:id:
    delete:
      consumes:
      - application/json
      description: 伪删除参数配置信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 参数配置伪删除
      tags:
      - 系统参数配置
    get:
      consumes:
      - application/json
      description: 详情参数配置信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.SysConfigResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 参数配置详情
      tags:
      - 系统参数配置
  /config/list:
    get:
      consumes:
      - application/json
      description: 获取所有参数配置信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.SysConfigResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 参数配置列表查询
      tags:
      - 系统参数配置
  /config/p:
    post:
      consumes:
      - application/json
      description: 获取所有参数配置信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.SysConfigResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 参数配置列表查询
      tags:
      - 系统参数配置
  /dept:
    post:
      consumes:
      - application/json
      description: 新增部门信息
      parameters:
      - description: 新增部门信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.DeptCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 部门新增
      tags:
      - 系统部门
    put:
      consumes:
      - application/json
      description: 更新部门信息
      parameters:
      - description: 更新部门信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.DeptUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 部门更新
      tags:
      - 系统部门
  /dept/:id:
    delete:
      consumes:
      - application/json
      description: 伪删除部门信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 部门伪删除
      tags:
      - 系统部门
    get:
      consumes:
      - application/json
      description: 详情部门信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.DeptResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 部门详情
      tags:
      - 系统部门
  /dept/list:
    get:
      consumes:
      - application/json
      description: 获取所有部门信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.DeptResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 部门列表查询
      tags:
      - 系统部门
  /dept/p:
    post:
      consumes:
      - application/json
      description: 获取所有部门信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.DeptResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 部门列表查询
      tags:
      - 系统部门
  /dept/tree:
    get:
      consumes:
      - application/json
      description: 获取所有部门信息(树形）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.DeptResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 部门列表查询(树形）
      tags:
      - 系统部门
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"